requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
httpx>=0.27.0
apscheduler>=3.10.4
